+++ src/Hy3Layout.hpp
@@ -113,6 +113,7 @@ public:
 	void shiftNode(Hy3Node&, ShiftDirection, bool once, bool visible);
 	void shiftWindow(const PHLWORKSPACE& workspace, ShiftDirection, bool once, bool visible);
 	void shiftFocus(const PHLWORKSPACE& workspace, ShiftDirection, bool visible, bool warp);
+	void warpCursor();
 	void moveNodeToWorkspace(const PHLWORKSPACE& origin, std::string wsname, bool follow);
 	void changeFocus(const PHLWORKSPACE& workspace, FocusShift);
 	void focusTab(